+++ NEWS.md
@@ -4,7 +4,7 @@
 
 * Add `log_odds_ratio` and `log_risk_ratio` function for inference.
 * Add `contrast_mat` to `treatment_effect` object.
-* Add `confint` to obtain confidence interval for `prediction_cf` object.
+* Add `confint` to obtain confidence interval for `prediction_cf` and `treatment_effect` object.
 
 # RobinCar2 0.1.1